+++ src/ArduinoJson/JsonVariantOr.hpp
@@ -6,6 +6,8 @@
 
 #include "Data/JsonVariantAs.hpp"
 #include "Polyfills/attributes.hpp"
+#include "TypeTraits/EnableIf.hpp"
+#include "TypeTraits/IsIntegral.hpp"
 
 namespace ArduinoJson {
 namespace Internals {
@@ -15,7 +17,8 @@ class JsonVariantOr {
  public:
   // Returns the default value if the JsonVariant is undefined of incompatible
   template <typename T>
-  T operator|(const T &defaultValue) const {
+  typename EnableIf<!IsIntegral<T>::value, T>::type operator|(
+      const T &defaultValue) const {
     if (impl()->template is<T>())
       return impl()->template as<T>();
     else
@@ -29,10 +32,21 @@ class JsonVariantOr {
     return value ? value : defaultValue;
   }
 
+  // Returns the default value if the JsonVariant is undefined of incompatible
+  // Special case for integers: we also accept double
+  template <typename Integer>
+  typename EnableIf<IsIntegral<Integer>::value, Integer>::type operator|(
+      const Integer &defaultValue) const {
+    if (impl()->template is<double>())
+      return impl()->template as<Integer>();
+    else
+      return defaultValue;
+  }
+
  private:
   const TImpl *impl() const {
     return static_cast<const TImpl *>(this);
   }
 };
-}
-}
+}  // namespace Internals
+}  // namespace ArduinoJson